Extended bitscope code to operate over UDP/IP
Tried it with syndey.bitscope.net and it didn't work - hung on 'T' command

Email from Bruce Tulloch (@bitscope.com) indicates that 'T' command is
no longer supported on newer bitscopes, which is why it didn't work.

Bruce suggested integrating with their BitLib library, but I think the
place for that it really COMEDI, so I'm ripping out bitscope support
for now.  May put it back in if COMEDI approach turns out to be unworkable.